Replace SFINAE with 'if constexpr' in any_traits.hpp
The complicated 'SFINAE' pattern for compile time code path
selection can be now be expressed much easier by the c++17
'if constexpr' in any_traits.hpp.

Since 'if constexpr' is only available in c++17 the required
(and used) c++ language standard is now set in the CMakeLists.txt.
Corresponding set statements were added in the CMakeLists.xpt
code generation template and the old compile flag '-std=c++11'
was removed from the same template file.

Additionally the dates of the copyright notice were updated
in License.xpt and CMakeLists.xpt.

All code (Ecore.ecore, emf4cpp.tests, XMLType) was regenerated
with the updated code generation templates.

The generated code of test 'EndUserLicense' was removed from
the repository because it is part of the test to
generate this code into a fresh (empty) folder.

Change-Id: I064ecbc7f7a7324b5bbad3d7a74c14c166a5a48f