Removal of deprecated std::iterator base class
Using std::iterator template as base class for iterators
is no longer recommended because it is marked as deprecated
from c++17 on and will issue compiler warnings if still used.

Therefore the use of std::iterator as base class was removed
and the corresponding typedefs are now defined directly in the
iterator classes.

Change-Id: I1aedaf6ff42aec55a9d926126f82d080955ac9dc