Support Alternative Serialization of EReferences
A href with empty URI (starting with '#') is now interpreted as
reference to an object in the same resource. This is in accordance
with the XMI standard (see https://www.omg.org/spec/XMI/2.5.1/PDF
page 64, row 2m. and 2n.).

Additional changes:

The XMLHandler.cpp has been re-formatted by clang-format.

Change-Id: Ie31fbb30227cc157fcfbddfe2a87d0d8d73123b0